start on moving away from doing modulecall and wasting ROM IDs
Now detecting the second ROM module by its identity string but not doing anything with that yet.
Detection works irrespective of module order. Numbering has been swapped in the id strings with the intention of swapping it around to be logical at some point. This will involve renaming the files etc to be consistent